Use the "composition over inheritance" technique
Instead of inheriting from `Array`, let's create an attribute that is an
array. The current object doesn't fit with the Liskov substitution
principle as, for instance, #sort is broken on Rubinius.

This makes the `PartsList` class compatible with Rubinius as `undef map`
will remove the map method but some other Enumerable methods (e.g. sort)
rely on it.